Add IndexOfAnyValues.Contains
