+++ src/libraries/System.Memory/tests/Span/LastIndexOfAny.T.cs
@@ -951,51 +951,5 @@ namespace System.SpanTests
                 }
             }
         }
-
-        [Fact]
-        [OuterLoop("Takes about a second to execute")]
-        public static void TestLastIndexOfAny_RandomInputs_Byte()
-        {
-            IndexOfAnyCharTestHelper.TestRandomInputs(
-                expected: LastIndexOfAnyReferenceImpl,
-                indexOfAny: (searchSpace, values) => searchSpace.LastIndexOfAny(values),
-                indexOfAnyValues: (searchSpace, values) => searchSpace.LastIndexOfAny(values));
-
-            static int LastIndexOfAnyReferenceImpl(ReadOnlySpan<byte> searchSpace, ReadOnlySpan<byte> values)
-            {
-                for (int i = searchSpace.Length - 1; i >= 0; i--)
-                {
-                    if (values.Contains(searchSpace[i]))
-                    {
-                        return i;
-                    }
-                }
-
-                return -1;
-            }
-        }
-
-        [Fact]
-        [OuterLoop("Takes about a second to execute")]
-        public static void TestLastIndexOfAny_RandomInputs_Char()
-        {
-            IndexOfAnyCharTestHelper.TestRandomInputs(
-                expected: LastIndexOfAnyReferenceImpl,
-                indexOfAny: (searchSpace, values) => searchSpace.LastIndexOfAny(values),
-                indexOfAnyValues: (searchSpace, values) => searchSpace.LastIndexOfAny(values));
-
-            static int LastIndexOfAnyReferenceImpl(ReadOnlySpan<char> searchSpace, ReadOnlySpan<char> values)
-            {
-                for (int i = searchSpace.Length - 1; i >= 0; i--)
-                {
-                    if (values.Contains(searchSpace[i]))
-                    {
-                        return i;
-                    }
-                }
-
-                return -1;
-            }
-        }
     }
 }
