+++ src/libraries/System.Private.CoreLib/src/System/IndexOfAnyValues/IndexOfAnyValuesInRange.cs
@@ -1,6 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Diagnostics;
 using System.Numerics;
 using System.Runtime.CompilerServices;
 
@@ -10,13 +11,19 @@ namespace System.Buffers
         where T : struct, INumber<T>
     {
         private readonly T _lowInclusive, _highInclusive;
+        private readonly uint _lowUint, _highMinusLow;
 
-        public IndexOfAnyValuesInRange(T lowInclusive, T highInclusive) =>
+        public IndexOfAnyValuesInRange(T lowInclusive, T highInclusive)
+        {
+            Debug.Assert(lowInclusive is byte or char);
             (_lowInclusive, _highInclusive) = (lowInclusive, highInclusive);
+            _lowUint = uint.CreateChecked(lowInclusive);
+            _highMinusLow = uint.CreateChecked(highInclusive - lowInclusive);
+        }
 
         internal override T[] GetValues()
         {
-            T[] values = new T[int.CreateChecked(_highInclusive - _lowInclusive)];
+            T[] values = new T[_highMinusLow + 1];
 
             T element = _lowInclusive;
             for (int i = 0; i < values.Length; i++)
@@ -27,6 +34,10 @@ namespace System.Buffers
 
             return values;
         }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal override bool ContainsCore(T value) =>
+            uint.CreateChecked(value) - _lowUint <= _highMinusLow;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal override int IndexOfAny(ReadOnlySpan<T> span) =>
