+++ src/libraries/System.Private.CoreLib/src/System/IndexOfAnyValues/IndexOfEmptyValues.cs
@@ -6,7 +6,11 @@ namespace System.Buffers
     internal sealed class IndexOfEmptyValues<T> : IndexOfAnyValues<T>
         where T : IEquatable<T>?
     {
-        internal override T[] GetValues() => Array.Empty<T>();
+        internal override T[] GetValues() =>
+            Array.Empty<T>();
+
+        internal override bool ContainsCore(T value) =>
+            false;
 
         internal override int IndexOfAny(ReadOnlySpan<T> span) =>
             -1;
