+++ src/libraries/System.Runtime/ref/System.Runtime.cs
@@ -7055,6 +7055,7 @@ namespace System.Buffers
     public class IndexOfAnyValues<T> where T : System.IEquatable<T>?
     {
         internal IndexOfAnyValues() { }
+        public bool Contains(T value) { throw null; }
     }
     public static class IndexOfAnyValues
     {
